+++ ui/src/features/common/confirm-modal/confirm-modal.tsx
@@ -3,13 +3,15 @@ import { Modal } from 'antd';
 import { ModalProps } from '../modal/use-modal';
 
 export interface ConfirmProps {
-  title: string;
+  title: string | React.ReactNode;
   onOk: () => void;
+  content?: string | React.ReactNode;
 }
 
 export const ConfirmModal = ({
   onOk,
   title = 'Are you sure?',
+  content,
   hide,
   visible
 }: ConfirmProps & ModalProps) => {
@@ -19,13 +21,8 @@ export const ConfirmModal = ({
   };
 
   return (
-    <Modal
-      closable={false}
-      open={visible}
-      onCancel={hide}
-      okText='Confirm'
-      onOk={onConfirm}
-      title={title}
-    />
+    <Modal open={visible} onCancel={hide} okText='Confirm' onOk={onConfirm} title={title}>
+      {content}
+    </Modal>
   );
 };
