+++ ui/src/features/project/credentials/types.ts
@@ -1,0 +1,2 @@
+export type CredentialsType = 'git' | 'helm' | 'image';
+export const CredentialTypeLabelKey = 'kargo.akuity.io/cred-type';
